+++ docs/scripts/views/drawer/dismissible-drawer.vue
@@ -11,7 +11,7 @@
           @click.native="() => { $router.back(); }">
           Back
         </ui-item-a>
-        <ui-divider></ui-divider>
+        <ui-list-divider></ui-list-divider>
       </ui-drawer-common-conntent>
     </ui-dismissible-drawer>