+++ src/scripts/components/divider.js
@@ -1,0 +1,7 @@
+import multiBootstrap from '../config/multi-bootstrap';
+// import UiDivider from './divider/divider';
+import UiTextDivider from './divider/text-divider';
+
+export default multiBootstrap({
+  UiTextDivider
+});