+++ src/scripts/components/divider/text-divider.vue
@@ -1,0 +1,39 @@
+<template>
+  <div :class="className">
+    <span>
+      <slot></slot>
+    </span>
+  </div>
+</template>
+
+<script>
+import getType from '../../utils/typeof';
+
+// Define constants
+const UI_TEXT_DIVIDER = {
+  TYPES: ['horizontal', 'vertical'],
+  VERTICAL: ['vertical', '|', 1]
+};
+
+export default {
+  name: 'ui-text-divider',
+  props: {
+    type: {
+      type: [String, Number],
+      default: UI_TEXT_DIVIDER.TYPES[0]
+    }
+  },
+  computed: {
+    isVertical() {
+      return UI_TEXT_DIVIDER.VERTICAL.includes(this.type);
+    },
+    className() {
+      return {
+        'mdc-text-divider': true,
+        'mdc-text-divider--horizontal': !this.isVertical,
+        'mdc-text-divider--vertical': this.isVertical
+      }
+    }
+  }
+};
+</script>