+++ src/scripts/components/list.js
@@ -7,8 +7,8 @@ import UiItemText from './list/item-text';
 import UiItemSubtext from './list/item-subtext';
 import UiListNav from './list/list-nav';
 import UiItemA from './list/item-a';
-import UiDivider from './divider/divider';
-import UiItemDivider from './divider/item-divider';
+import UiListDivider from './list/list-divider';
+import UiItemDivider from './list/item-divider';
 
 export default multiBootstrap({
   UiListGroup,
@@ -19,6 +19,6 @@ export default multiBootstrap({
   UiItemSubtext,
   UiListNav,
   UiItemA,
-  UiDivider,
+  UiListDivider,
   UiItemDivider
 });